Change the delegated performance APIs to have a single setter.
Summary:
This improves Swift usability.

Properties in protocols have to be nullable, otherwise the initializer will be forced to assign some non-null value to the object. This means that swift performers have to unbox the nullable type every time they invoke one of the properties. This is not ideal.

Ideally, the properties would be nonnull. Doing so would require that the protocol also include an initializer that can initialize the properties. Increasing the number of arguments to the designated initializer in this manner isn't scalable and could likely force us into a situation where two protocols clash.

Based on the above, this diff takes the approach of using a single setter to provide both of the delegated performance blocks. The performer object is responsible for storing the blocks in some manner.

This new API is being introduced as an optional API in order to not cause a breaking release. In the next release we will delete the deprecated APIs and make the setter required. This subsequent release will be a major release.

Reviewers: O4 Material Motion Apple platform reviewers!, markwei, O2 Material Motion

Reviewed By: markwei, O2 Material Motion

Subscribers: markwei

Tags: #material_motion

Differential Revision: http://codereview.cc/D1307