MDMPlan now formally conforms to NSCopying.
Summary:
We copy plans when they are committed to a scheduler so that plan data is cleanly provided to the performer without needing to worry about subsequent changes to the original plans affecting the performer after the fact.

Includes a test confirming that plans are indeed copied. This test required a new plans created trace notification.

Reviewers: O4 Material Motion Apple platform reviewers, O2 Material Motion, markwei

Reviewed By: O4 Material Motion Apple platform reviewers, O2 Material Motion, markwei

Tags: #material_motion

Differential Revision: http://codereview.cc/D1601