Add Life of a Configurable Plan example.
Summary: This example demonstrates how to configure a performer using properties on a plan. We recreate the Draggable plan found in Life of a Plan, but this time with a configurable panGestureRecognizer property. We assign a pre-made pan gesture recognizer that's already associated with the view controller's root view, allowing us to drag the square by touching anywhere in the view controller.

Reviewers: O4 Material Motion Apple platform reviewers, O2 Material Motion, markwei

Reviewed By: O4 Material Motion Apple platform reviewers, O2 Material Motion, markwei

Tags: #material_motion

Differential Revision: http://codereview.cc/D1603