Change Swift performer addPlan API to match latest conventions.
Summary: This is unfortunately a breaking change for all Swift performers. I do not see a clear path forward for supporting both the old API and the new because I do not know of a way to mark only the Swift version of the API name as deprecated.

Reviewers: O2 Material Motion, O4 Material Motion Apple platform reviewers!

Tags: #material_motion

Differential Revision: http://codereview.cc/D1718