Simplify the internal performer group logic.
Summary:
Prior to this change we used performer info objects to track performer + activity state. This created a fairly complex tree of of objects to maintain (`Runtime => PerformerGroups => PerformerInfo => Performer`). This change removes the performer info object so that the internals are easier to understand. As part of this change we've moved the is-active token pool up a level into the performer group.

Specific changes:

- Replace token generator cross dependencies with delegate pattern.
- Remove performer info object.
- Token activity is now tracked per target. This change is in preparation to make tokens tracked across the runtime as a global pool. This will make it easier to build token group APIs for observing subset activity changes.

Reviewers: O4 Material Motion Apple platform reviewers, O2 Material Motion, markwei

Reviewed By: O4 Material Motion Apple platform reviewers, O2 Material Motion, markwei

Tags: #material_motion

Maniphest Tasks: T89

Differential Revision: http://codereview.cc/D2013