Move runtime target registry logic out to a separate class.
Summary:
This diff is in preparation for supporting recursive activity groups. The runtime will be the root activity group. Activity groups can create child activity groups. Plans can be added to any activity group, and these plans are ultimately added to the same target registry for a given runtime. In order to keep a clean separation of concerns, this diff introduces a separate target registry object that will facilitate all performer creation and ownership. In a follow-up diff, activity groups will make use of this shared object to add plans.

Depends on D2038.

Reviewers: markwei, O2 Material Motion, O4 Material Motion Apple platform reviewers

Reviewed By: markwei, O2 Material Motion, O4 Material Motion Apple platform reviewers

Subscribers: markwei

Tags: #material_motion

Differential Revision: http://codereview.cc/D2040