[breaking] Tokens are now persisted in a shared pool and expected to be reused.
Summary:
Rather than generate and terminate a token each time work begins/ends, a performer is now expected to fetch a token for a given plan and to set the isActive property on the token as the associated work's state changes. This new API makes it easier to build a system of recursive activity groups that intend to watch activity state changes for a given collection of plans. It is easier because there is now a guaranteed 1-1 mapping of plans to tokens, meaning an entity can subscribe to changes in token's (i.e. the plan's) state.

Notable changes:

- Runtime adds itself as an activity observer for a plan's token before the plan is committed to the target registry.
- Performers now receive a **plan tokenizer**, not a token generator. This is a more accurate reflection of the fact that this object turns plans into their corresponding token.
- Plan emitters now store a reference to the TargetRegistry rather than to the runtime.
- Plan emitter still needs to access the runtime in order to ensure that token observation is set up when new plans are emitted. This will be resolved in a follow-up diff.

Depends on D2040.

Reviewers: O4 Material Motion Apple platform reviewers, O2 Material Motion, markwei

Reviewed By: O4 Material Motion Apple platform reviewers, O2 Material Motion, markwei

Subscribers: markwei

Tags: #material_motion

Differential Revision: http://codereview.cc/D2041